Legal-Contract-Analysis-and-Risk-Assessment-Bot-main/requirements.txt
streamlit
plotly
langdetect
python-dotenv
google-generativeai
pypdfium2
python-docx
fpdf